--- src/app/hooks/use-modal-hooks.ts
@@ -1,10 +1,0 @@
-import { useState } from "react";
-
-export const useModal = () => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const openModal = () => setIsOpen(true);
-  const closeModal = () => setIsOpen(false);
-
-  return { isOpen, openModal, closeModal };
-};
